# Reduce PostingsForMatchersCache.expire() pressure on mutex