@pracucci 2024-10-29
> Changes LGTM. It seems like this is a case where `RWLock.TryRLock()` could be used but I don't feel strongly about using it instead of the atomic here.

I tried it, but it looks slower. Here is the comparison between this PR vs `Try...Lock`:

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Apple M3 Pro
                                                          │ before.txt  │              after.txt               │
                                                          │   sec/op    │   sec/op     vs base                 │
PostingsForMatchersCache_ConcurrencyOnHighEvictionRate-11   303.8n ± 2%   795.8n ± 2%  +161.93% (p=0.000 n=10)

                                                          │ before.txt  │              after.txt              │
                                                          │    B/op     │    B/op      vs base                │
PostingsForMatchersCache_ConcurrencyOnHighEvictionRate-11   1016.0 ± 0%   1371.5 ± 0%  +34.99% (p=0.000 n=10)

                                                          │ before.txt │             after.txt              │
                                                          │ allocs/op  │ allocs/op   vs base                │
PostingsForMatchersCache_ConcurrencyOnHighEvictionRate-11   21.00 ± 0%   27.00 ± 0%  +28.57% (p=0.000 n=10)
```

Note: it's still faster than `main`, but not as fast as this PR.