@56quarters 2024-10-29
Makes sense, thanks!